requirements.txt
datasets==3.2.0
numpy==2.2.1
openai==1.59.3
Pillow==11.1.0
torch==2.5.1
transformers==4.47.1
